+++ src/test/java/com/renomad/minum/logging/CustomLoggingLevel.java
@@ -1,0 +1,9 @@
+package com.renomad.minum.logging;
+
+public enum CustomLoggingLevel {
+
+    /**
+     * For logging each request as it comes in
+     */
+    REQUEST
+}